migrate files fixed and new columns added to user table

+++ db/migrate/20170918202127_drop_table_materia.rb
@@ -1,5 +1,5 @@
 class DropTableMateria < ActiveRecord::Migration[5.1]
   def change
-    drop_table :materia
+    drop_table :materia if ActiveRecord::Base.connection.table_exists? 'materia'
   end
 end
